chore: reorgranize helper functions
