+++ composer.json
@@ -35,7 +35,9 @@
             "Give\\": "src/"
         },
         "files": [
-            "src/Helpers/Form/Theme.php",
+            "src/Helpers/Form/Theme/Theme.php",
+            "src/Helpers/Form/Theme/Utils/Admin.php",
+            "src/Helpers/Form/Theme/Utils/Frontend.php",
             "src/Helpers/Form/Utils.php"
         ]
     }
